Moving common functions to new file
