+++ integration-test/step-definitions/processes_all_messages_if_there_is_no_error.steps.ts
@@ -2,27 +2,12 @@ import { SQS } from 'aws-sdk';
 import { defineFeature, loadFeature } from 'jest-cucumber';
 
 import { SQSConsumer } from '../..';
+import AWSHelper from '../AWSHelper';
 
 const feature = loadFeature('./integration-test/features/processes_all_messages_if_there_is_no_error.feature');
 
-async function getTotalNumberOfMessagesInQueue(options: {
-  sqs: SQS;
-  queueUrl: string;
-}): Promise<number> {
-  const sqsQueueAttributes: SQS.GetQueueAttributesResult = await options.sqs.getQueueAttributes({
-    QueueUrl: options.queueUrl,
-    AttributeNames: [
-      'All',
-    ],
-  }).promise();
-  expect(sqsQueueAttributes.Attributes).toBeDefined();
-  const numberOfMessagesVisible: number = parseInt(sqsQueueAttributes.Attributes!.ApproximateNumberOfMessages, 10);
-  const numberOfMessagesDelayed: number = parseInt(sqsQueueAttributes.Attributes!.ApproximateNumberOfMessagesDelayed, 10);
-  const numberOfMessagesNotVisible: number = parseInt(sqsQueueAttributes.Attributes!.ApproximateNumberOfMessagesNotVisible, 10);
-
-  return (numberOfMessagesVisible + numberOfMessagesDelayed + numberOfMessagesNotVisible);
-}
 jest.setTimeout(1200000);
+const awsHelper: AWSHelper = new AWSHelper();
 defineFeature(feature, test => {
   test('When there are no errors, processes all messages', ({ given, when, then }) => {
     const sqs: SQS = new SQS({
@@ -46,12 +31,9 @@ defineFeature(feature, test => {
       console.log(message);
     }));
     given('there are messages in sqs queue', async (): Promise<void> => {
-      await sqs.purgeQueue({
-        QueueUrl: queueUrl,
-      }).promise();
-      await sqs.purgeQueue({
-        QueueUrl: dlqUrl,
-      }).promise();
+      await awsHelper.purgeQueues({
+        sqs: sqs,
+      });
       Atomics.wait(
         new Int32Array(new SharedArrayBuffer(4)), 0, 0, 1000
       );
@@ -109,11 +91,11 @@ defineFeature(feature, test => {
     });
 
     then('all messages should be processed without any messages going to DLQ', async (): Promise<void> => {
-      const totalMessagesInMainQueue: number = await getTotalNumberOfMessagesInQueue({
+      const totalMessagesInMainQueue: number = await awsHelper.getTotalNumberOfMessagesInQueue({
         sqs: sqs,
         queueUrl: queueUrl,
       });
-      const totalMessagesInDLQ: number = await getTotalNumberOfMessagesInQueue({
+      const totalMessagesInDLQ: number = await awsHelper.getTotalNumberOfMessagesInQueue({
         sqs: sqs,
         queueUrl: dlqUrl,
       });
